requirements.txt
langchain
langgraph
duckduckgo-search
chromadb
openai
tiktoken
python-dotenv
streamlit
requests
